Break up utils into separate files
