+++ dbt_copilot_helper/commands/waf.py
@@ -11,11 +11,11 @@ import yaml
 from cfn_tools import load_yaml
 
 from dbt_copilot_helper.commands.dns import get_load_balancer_domain_and_configuration
-from dbt_copilot_helper.utils import ClickDocOptGroup
-from dbt_copilot_helper.utils import check_aws_conn
-from dbt_copilot_helper.utils import check_response
-from dbt_copilot_helper.utils import ensure_cwd_is_repo_root
-from dbt_copilot_helper.utils import get_lint_result
+from dbt_copilot_helper.utils.aws import check_aws_conn
+from dbt_copilot_helper.utils.aws import check_response
+from dbt_copilot_helper.utils.click import ClickDocOptGroup
+from dbt_copilot_helper.utils.cloudformation import get_lint_result
+from dbt_copilot_helper.utils.files import ensure_cwd_is_repo_root
 
 # This may need to change, once we determine what the default WAF name will be.
 WAF_DEFAULT_NAME = "default"
