+++ dbt_copilot_helper/utils/files.py
@@ -1,0 +1,36 @@
+from pathlib import Path
+
+import click
+
+
+def mkdir(base, path):
+    if (base / path).exists():
+        return f"Directory {path} exists; doing nothing"
+
+    (base / path).mkdir(parents=True)
+    return f"Directory {path} created"
+
+
+def mkfile(base, path, contents, overwrite=False):
+    file_exists = (base / path).exists()
+
+    if file_exists and not overwrite:
+        return f"File {path} exists; doing nothing"
+
+    action = "overwritten" if overwrite else "created"
+
+    with open(base / path, "w") as fd:
+        fd.write(contents)
+
+    return f"File {path} {action}"
+
+
+def ensure_cwd_is_repo_root():
+    """Exit if we're not in the root of the repo."""
+
+    if not Path("./copilot").exists() or not Path("./copilot").is_dir():
+        click.secho(
+            "Cannot find copilot directory. Run this command in the root of the deployment repository.",
+            bg="red",
+        )
+        exit(1)
