+++ tests/utils/test_validation.py
@@ -1,0 +1,24 @@
+import pytest
+from schema import SchemaError
+
+from dbt_copilot_helper.utils.validation import validate_string
+
+
+@pytest.mark.parametrize(
+    "regex_pattern, valid_string, invalid_string",
+    [(r"^\d+-\d+$", "1-10", "20-21-23"), (r"^\d+s$", "10s", "10seconds")],
+)
+def test_validate_string(regex_pattern, valid_string, invalid_string):
+    validator = validate_string(regex_pattern)
+
+    assert validator(valid_string) == valid_string
+
+    with pytest.raises(SchemaError) as err:
+        validator(invalid_string)
+
+    assert (
+        err.value.args[0]
+        == f"String '{invalid_string}' does not match the required pattern '{regex_pattern}'. For "
+           "more details on valid string patterns see: "
+           "https://aws.github.io/copilot-cli/docs/manifest/lb-web-service/"
+    )
